remove mock/QA, mock/brickinfo, mock/sample
